requirements.txt
qiskit
qiskit-aer
numpy
customtkinter
pillow
matplotlib
pylatexenc
